+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## 1.0.0
+* TDL-22064 add integration tests [#10](https://github.com/singer-io/tap-circle-ci/pull/10)
+* Fix Catalog/Stream Metadata issues [#13](https://github.com/singer-io/tap-circle-ci/pull/13)
+* TDL-22062 api client upgrade [#14](https://github.com/singer-io/tap-circle-ci/pull/14)
+* TDL-22066 Schema Changes [#15](https://github.com/singer-io/tap-circle-ci/pull/15)
+* TDL-22353 class based implementation [#16](https://github.com/singer-io/tap-circle-ci/pull/16)
 
 ## 0.1.2
   * Fix a bookmark bug [#3](https://github.com/singer-io/tap-circle-ci/pull/3/)